change uint8 to float32 to reduce mean works.

+++ src/conv_pool_net2.py
@@ -24,7 +24,7 @@ def conv_conv_conv_pool(x, chan_in, chan_out, name):
   tf.summary.histogram(name+'_conv1', relu1)
   tf.summary.scalar(name+'ssq_kernel1', tf.reduce_sum(tf.square(kernel)))
   tf.summary.scalar(name + 'dead_kernel1', tf.reduce_mean(
-    tf.cast(tf.less(conv1+b, 0), tf.uint8)))
+    tf.cast(tf.less(conv1+b, 0), tf.float32)))
 
 
   with tf.variable_scope(name+'_conv2',
@@ -39,7 +39,7 @@ def conv_conv_conv_pool(x, chan_in, chan_out, name):
   tf.summary.histogram(name+'_conv2', relu2)
   tf.summary.scalar(name + 'ssq_kernel2', tf.reduce_sum(tf.square(kernel)))
   tf.summary.scalar(name + 'dead_kernel2', tf.reduce_mean(
-    tf.cast(tf.less(conv2 + b, 0), tf.uint8)))
+    tf.cast(tf.less(conv2 + b, 0), tf.float32)))
 
   with tf.variable_scope(name+'_conv3',
                          initializer=tf.truncated_normal_initializer(stddev=(2.0 /
@@ -53,7 +53,7 @@ def conv_conv_conv_pool(x, chan_in, chan_out, name):
   tf.summary.histogram(name+'_conv3', relu3)
   tf.summary.scalar(name + 'ssq_kernel3', tf.reduce_sum(tf.square(kernel)))
   tf.summary.scalar(name + 'dead_kernel3', tf.reduce_mean(
-    tf.cast(tf.less(conv3 + b, 0), tf.uint8)))
+    tf.cast(tf.less(conv3 + b, 0), tf.float32)))
 
   y = tf.nn.max_pool(relu3, [1, 2, 2, 1], [1, 2, 2, 1], padding='SAME')
   tf.summary.histogram(name+'_maxpool', y)
@@ -74,7 +74,7 @@ def conv_conv_pool(x, chan_in, chan_out, name):
   tf.summary.histogram(name+'_conv1', relu1)
   tf.summary.scalar(name + 'ssq_kernel1', tf.reduce_sum(tf.square(kernel)))
   tf.summary.scalar(name + 'dead_kernel1', tf.reduce_mean(
-    tf.cast(tf.less(conv1 + b, 0), tf.uint8)))
+    tf.cast(tf.less(conv1 + b, 0), tf.float32)))
 
   with tf.variable_scope(name+'_conv2',
                          initializer=tf.truncated_normal_initializer(stddev=(2.0 /
@@ -88,7 +88,7 @@ def conv_conv_pool(x, chan_in, chan_out, name):
   tf.summary.histogram(name+'_conv2', relu2)
   tf.summary.scalar(name + 'ssq_kernel2', tf.reduce_sum(tf.square(kernel)))
   tf.summary.scalar(name + 'dead_kernel2', tf.reduce_mean(
-    tf.cast(tf.less(conv2 + b, 0), tf.uint8)))
+    tf.cast(tf.less(conv2 + b, 0), tf.float32)))
 
   y = tf.nn.max_pool(relu2, [1, 2, 2, 1], [1, 2, 2, 1], padding='SAME')
   tf.summary.histogram(name+'_maxpool', y)
@@ -129,7 +129,7 @@ def conv_pool_net2(training_batch, config):
   tf.summary.histogram('fc1', fc1)
   tf.summary.scalar('ssq_fc1', tf.reduce_sum(tf.square(wfc1)))
   tf.summary.scalar('dead_fc1', tf.reduce_mean(
-    tf.cast(tf.less(tf.matmul(flat1, wfc1) + bfc1, 0), tf.uint8)))
+    tf.cast(tf.less(tf.matmul(flat1, wfc1) + bfc1, 0), tf.float32)))
   if config.dropout:
     fc1 = tf.nn.dropout(fc1, config.dropout_keep_prob)
 
@@ -145,7 +145,7 @@ def conv_pool_net2(training_batch, config):
   tf.summary.histogram('fc2', fc2)
   tf.summary.scalar('ssq_fc2', tf.reduce_sum(tf.square(wfc2)))
   tf.summary.scalar('dead_fc2', tf.reduce_mean(
-    tf.cast(tf.less(tf.matmul(fc1, wfc2) + bfc2, 0), tf.uint8)))
+    tf.cast(tf.less(tf.matmul(fc1, wfc2) + bfc2, 0), tf.float32)))
   if config.dropout:
     fc2 = tf.nn.dropout(fc2, config.dropout_keep_prob)
 
@@ -161,7 +161,7 @@ def conv_pool_net2(training_batch, config):
   tf.summary.histogram('fc3', fc3)
   tf.summary.scalar('ssq_fc3', tf.reduce_sum(tf.square(wfc3)))
   tf.summary.scalar('dead_fc3', tf.reduce_mean(
-    tf.cast(tf.less(tf.matmul(fc2, wfc3) + bfc3, 0), tf.uint8)))
+    tf.cast(tf.less(tf.matmul(fc2, wfc3) + bfc3, 0), tf.float32)))
   if config.dropout:
     fc3 = tf.nn.dropout(fc3, config.dropout_keep_prob)
